requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-SocketIO
eventlet
python-dotenv
Werkzeug
